pyproject.toml
[project]
name = "anime"
version = "0.1.0"
description = "处理弹幕"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opencc-python-reimplemented>=0.1.7",
    "requests>=2.32.3",
]
